make installing equal versions a no-op
Here we introduce a "non-error" `sm.ErrNoAction` to signal to the task
handler that there is nothing more to do (as though there were a real
error condition), but that the task handler should continue processing
additional firmware install tasks.

The result of this change is that attempting to install firmware at a
version already installed will not fail the task. This should avoid the
need to force install a firmware set which may overwrite firmware
already at the correct version and thus cost more time to apply a
condition.