outofband/action_handlers: poller to collect inventory for install tasks
This fixes the poller failure when comparing the installed firmware with
the expected for BMC firmware - for firmware upload verify tasks.

The poller is invoked for both install and firmware upload verify tasks,
it did not differentiate between the two before attempting to collect
inventory from the BMC to verify the installed firmware matches the
expected. When verifying a firmware upload task, an inventory and
version comparision is not required.